backend/requirements.txt
fastapi
uvicorn[standard]
litellm
openai
qdrant-client
psycopg[binary]
sqlalchemy
pydantic
pydantic-settings
python-dotenv
cohere
tiktoken
python-multipart
anthropic
pyyaml
